Fix Linux CI build errors on integration (round 3)
Third batch surfaced behind round 2 (-Werror cascade), all in merged features:

- ChallengeMode.h: include Common.h for IN_MILLISECONDS (used by the death /
  affix timing constants) — was undeclared in that translation unit.
- Player.cpp: drop the unused named binding in the gossip-spell cast
  (if (SpellInfo const* spellInfo = ...) -> if (sSpellMgr->GetSpellInfo(...)));
  the body casts via item->SpellID and never referenced spellInfo
  (-Werror=unused-variable).
- GarrisonMgr::GetFollowerZoneSupportSpell: cast the int32 GarrFollowerID to
  uint32 to match the garrFollowerID parameter (-Werror=sign-compare).

+++ src/server/game/Garrison/GarrisonMgr.cpp
@@ -458,7 +458,7 @@ uint32 GarrisonMgr::GetFollowerZoneSupportSpell(uint32 garrFollowerID, uint32 fa
     GarrFollSupportSpellEntry const* best = nullptr;
     for (GarrFollSupportSpellEntry const* entry : sGarrFollSupportSpellStore)
     {
-        if (entry->GarrFollowerID != garrFollowerID)
+        if (uint32(entry->GarrFollowerID) != garrFollowerID)
             continue;
         if (!best || entry->OrderIndex < best->OrderIndex)
             best = entry;
